backend_ml/requirements.txt
pandas>=2.2.3
requests==2.31.0
pymongo==4.6.1
openpyxl==3.1.2
python-dotenv==1.0.0
tqdm==4.66.1
fastapi==0.104.1
uvicorn==0.24.0
meson-python
ninja
cython
numpy>=1.26.0
